requirements.txt
asgiref==3.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
Django==4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
importlib_resources==6.4.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pkgutil_resolve_name==1.3.10
psycopg2==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
zipp==3.18.1
